Factor common IAM policy bits into 'google.cloud.iam'.
Pubsub-specific roles, permissions left behind in 'google.cloud.pubsub.iam'.

'google.cloud.pubsub.iam.Policy' subclasses the core one, extending it
to deal with the pubsub-specific roles.
