+++ pubsub/tests/unit/test_iam.py
@@ -48,111 +48,33 @@ class TestPolicy(unittest.TestCase):
         self.assertEqual(list(policy.publishers), [])
         self.assertEqual(list(policy.subscribers), [])
 
-    def test_user(self):
-        EMAIL = 'phred@example.com'
-        MEMBER = 'user:%s' % (EMAIL,)
-        policy = self._make_one()
-        self.assertEqual(policy.user(EMAIL), MEMBER)
-
-    def test_service_account(self):
-        EMAIL = 'phred@example.com'
-        MEMBER = 'serviceAccount:%s' % (EMAIL,)
-        policy = self._make_one()
-        self.assertEqual(policy.service_account(EMAIL), MEMBER)
-
-    def test_group(self):
-        EMAIL = 'phred@example.com'
-        MEMBER = 'group:%s' % (EMAIL,)
-        policy = self._make_one()
-        self.assertEqual(policy.group(EMAIL), MEMBER)
-
-    def test_domain(self):
-        DOMAIN = 'example.com'
-        MEMBER = 'domain:%s' % (DOMAIN,)
-        policy = self._make_one()
-        self.assertEqual(policy.domain(DOMAIN), MEMBER)
-
-    def test_all_users(self):
-        policy = self._make_one()
-        self.assertEqual(policy.all_users(), 'allUsers')
-
-    def test_authenticated_users(self):
-        policy = self._make_one()
-        self.assertEqual(policy.authenticated_users(), 'allAuthenticatedUsers')
-
-    def test_from_api_repr_only_etag(self):
-        RESOURCE = {
-            'etag': 'ACAB',
-        }
-        klass = self._get_target_class()
-        policy = klass.from_api_repr(RESOURCE)
-        self.assertEqual(policy.etag, 'ACAB')
-        self.assertIsNone(policy.version)
-        self.assertEqual(list(policy.owners), [])
-        self.assertEqual(list(policy.editors), [])
-        self.assertEqual(list(policy.viewers), [])
-
-    def test_from_api_repr_complete(self):
+    def test__bind_custom_role_publisher(self):
         from google.cloud.pubsub.iam import (
-            OWNER_ROLE,
-            EDITOR_ROLE,
-            VIEWER_ROLE,
             PUBSUB_PUBLISHER_ROLE,
+        )
+        PUBLISHER = 'user:phred@example.com'
+        policy = self._make_one()
+        policy._bind_custom_role(PUBSUB_PUBLISHER_ROLE, set([PUBLISHER]))
+
+        self.assertEqual(sorted(policy.publishers), [PUBLISHER])
+
+    def test__bind_custom_role_subscriber(self):
+        from google.cloud.pubsub.iam import (
             PUBSUB_SUBSCRIBER_ROLE,
         )
-
-        OWNER1 = 'user:phred@example.com'
-        OWNER2 = 'group:cloud-logs@google.com'
-        EDITOR1 = 'domain:google.com'
-        EDITOR2 = 'user:phred@example.com'
-        VIEWER1 = 'serviceAccount:1234-abcdef@service.example.com'
-        VIEWER2 = 'user:phred@example.com'
-        PUBLISHER = 'user:phred@example.com'
         SUBSCRIBER = 'serviceAccount:1234-abcdef@service.example.com'
-        RESOURCE = {
-            'etag': 'DEADBEEF',
-            'version': 17,
-            'bindings': [
-                {'role': OWNER_ROLE, 'members': [OWNER1, OWNER2]},
-                {'role': EDITOR_ROLE, 'members': [EDITOR1, EDITOR2]},
-                {'role': VIEWER_ROLE, 'members': [VIEWER1, VIEWER2]},
-                {'role': PUBSUB_PUBLISHER_ROLE, 'members': [PUBLISHER]},
-                {'role': PUBSUB_SUBSCRIBER_ROLE, 'members': [SUBSCRIBER]},
-            ],
-        }
-        klass = self._get_target_class()
-        policy = klass.from_api_repr(RESOURCE)
-        self.assertEqual(policy.etag, 'DEADBEEF')
-        self.assertEqual(policy.version, 17)
-        self.assertEqual(sorted(policy.owners), [OWNER2, OWNER1])
-        self.assertEqual(sorted(policy.editors), [EDITOR1, EDITOR2])
-        self.assertEqual(sorted(policy.viewers), [VIEWER1, VIEWER2])
-        self.assertEqual(sorted(policy.publishers), [PUBLISHER])
+        policy = self._make_one()
+        policy._bind_custom_role(PUBSUB_SUBSCRIBER_ROLE, set([SUBSCRIBER]))
+
         self.assertEqual(sorted(policy.subscribers), [SUBSCRIBER])
 
-    def test_from_api_repr_bad_role(self):
-        BOGUS1 = 'user:phred@example.com'
-        BOGUS2 = 'group:cloud-logs@google.com'
-        RESOURCE = {
-            'etag': 'DEADBEEF',
-            'version': 17,
-            'bindings': [
-                {'role': 'nonesuch', 'members': [BOGUS1, BOGUS2]},
-            ],
-        }
-        klass = self._get_target_class()
-        with self.assertRaises(ValueError):
-            klass.from_api_repr(RESOURCE)
-
-    def test_to_api_repr_defaults(self):
+    def test__bind_custom_role_unknown(self):
         policy = self._make_one()
-        self.assertEqual(policy.to_api_repr(), {})
+        USER = 'user:phred@example.com'
+        with self.assertRaises(ValueError):
+            policy._bind_custom_role('nonesuch', set([USER]))
 
-    def test_to_api_repr_only_etag(self):
-        policy = self._make_one('DEADBEEF')
-        self.assertEqual(policy.to_api_repr(), {'etag': 'DEADBEEF'})
-
-    def test_to_api_repr_full(self):
+    def test__role_bindings(self):
         from google.cloud.pubsub.iam import (
             PUBSUB_ADMIN_ROLE,
             PUBSUB_EDITOR_ROLE,
@@ -169,17 +91,13 @@ class TestPolicy(unittest.TestCase):
         VIEWER2 = 'user:phred@example.com'
         PUBLISHER = 'user:phred@example.com'
         SUBSCRIBER = 'serviceAccount:1234-abcdef@service.example.com'
-        EXPECTED = {
-            'etag': 'DEADBEEF',
-            'version': 17,
-            'bindings': [
-                {'role': PUBSUB_ADMIN_ROLE, 'members': [OWNER1, OWNER2]},
-                {'role': PUBSUB_EDITOR_ROLE, 'members': [EDITOR1, EDITOR2]},
-                {'role': PUBSUB_VIEWER_ROLE, 'members': [VIEWER1, VIEWER2]},
-                {'role': PUBSUB_PUBLISHER_ROLE, 'members': [PUBLISHER]},
-                {'role': PUBSUB_SUBSCRIBER_ROLE, 'members': [SUBSCRIBER]},
-            ],
-        }
+        EXPECTED = [
+            {'role': PUBSUB_ADMIN_ROLE, 'members': [OWNER1, OWNER2]},
+            {'role': PUBSUB_EDITOR_ROLE, 'members': [EDITOR1, EDITOR2]},
+            {'role': PUBSUB_VIEWER_ROLE, 'members': [VIEWER1, VIEWER2]},
+            {'role': PUBSUB_PUBLISHER_ROLE, 'members': [PUBLISHER]},
+            {'role': PUBSUB_SUBSCRIBER_ROLE, 'members': [SUBSCRIBER]},
+        ]
         policy = self._make_one('DEADBEEF', 17)
         policy.owners.add(OWNER1)
         policy.owners.add(OWNER2)
@@ -189,4 +107,4 @@ class TestPolicy(unittest.TestCase):
         policy.viewers.add(VIEWER2)
         policy.publishers.add(PUBLISHER)
         policy.subscribers.add(SUBSCRIBER)
-        self.assertEqual(policy.to_api_repr(), EXPECTED)
+        self.assertEqual(policy._role_bindings(), EXPECTED)
